first step

+++ backend2/src/main/java/com/netcracker/edu/backend2/repository/SubscriptionRepository.java
@@ -1,12 +1,11 @@
 package com.netcracker.edu.backend2.repository;
 
 import com.netcracker.edu.backend2.entities.SubscriptionsEntity;
-import com.netcracker.edu.backend2.entities.UsersEntity;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface SubscriptionRepository extends CrudRepository<SubscriptionsEntity, Long> {
+public interface SubscriptionRepository extends CrudRepository<SubscriptionsEntity, Integer> {
     List<SubscriptionsEntity> findSubscriptionsEntitiesByUsersByUserId_UserId(int id);
 }
